# specs/001-telco-adb-automation/contracts/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telco ADB Automation API",
    "description": "REST API for managing devices, flows, and executions in the Telco ADB Automation Desktop Framework",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/devices/{device_id}": {
      "get": {
        "summary": "Get device details",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Device details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Device" }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        }
      }
    },
    "/modules": {
      "get": {
        "summary": "List available modules",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "active_only",
            "in": "query",
            "schema": { "type": "boolean", "default": true }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available modules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Module" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Device": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "phone_number": { "type": "string", "nullable": true },
          "sim_info": { "type": "object" },
          "model": { "type": "string" },
          "os_version": { "type": "string" },
          "status": { "type": "string", "enum": ["connected", "disconnected", "busy", "error"] },
          "last_seen": { "type": "string", "format": "date-time" },
          "capabilities": { "type": "object" }
        }
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "version": { "type": "string" },
          "category": { "type": "string" },
          "steps": { "type": "array", "items": { "type": "string" } },
          "input_schema": { "type": "object" },
          "timeout_seconds": { "type": "integer" },
          "requires_root": { "type": "boolean" }
        }
      },
...
    }
  }
}